requirements.txt
Django==2.0.4
django-crispy-forms==1.7.2
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Markdown==2.6.11
Pillow==5.3.0
PyJWT==1.6.4
pytz==2018.3
